refactor: Drive control logic into new file - drive_control.hpp
